refactor(vaultcore): extract Typer-free edit engine, thin CLI renderer
Move the body/frontmatter edit pipeline (_resolve_doc_path, _split_document,
_enforce_blob_hash, _compose_new_text, _validate_proposed, _write_proposed,
_EditError) out of cli/edit_cmd.py into a new vaultcore.edit_engine module,
adding a result-returning execute_edit core and a typed EditResult dataclass
with no Typer, console, or typer.Exit coupling. Re-point cmd_set_body,
cmd_set_frontmatter, and cmd_edit as thin renderers over execute_edit,
mirroring status_cmd over vaultcore.orientation. Behavior, envelopes, and
exit codes preserved; existing edit and rename CLI tests stay green.

Implements P01.S01 and P01.S02 of 2026-07-09-mcp-tool-schema-plan.
